backend/requirements.txt
Flask
Flask-Cors
Pillow
torch
diffusers
transformers
accelerate
safetensors